+++ CHANGELOG.md
@@ -3,8 +3,7 @@
 ## Version ?? UNKNOWN ??
 
 * **New feature:** `!voteTimer 5 minutes` to start a countdown timer that will perform the `!vote` command after 5 minutes.</br>Times can range from 15 seconds to 20 minutes.</br>Also has corresponding `!stopVoteTimer` command to cancel the timer, as well as shortcuts `!vt` and `!svt`.
-* **New feature:** `The bot can now look up roles with the `!role` command - this requires some setup:
-  * `!addset <json url>` adds a custom set to the list for this server
+* **New feature:** The bot can now look up roles with the `!role` command.</br>For custom roles this requires some setup:</br>`!addScript <json url>` adds a custom script to the list of scripts for this server. Roles from that script are available via the `!role` command.
 * Added `!setChatChannel <chat channel name>` command to support `!voteTimer`.
 
 ## Version 1.1.0